Organizing and extending tests - include async tests
